Fix desired capabilities in Undetected Chromedriver for Selenium 4.10
The desired_capabilities keyword has been deprecated for a while in
Selenium and was officially removed in Selenium 4.10. Therefore, using
the latest version of Undetected Chromedriver would accept the
desired_capabilities keyword, but they wouldn't actually be applied to
the webdriver.

This commit applies the capabilities in the manner that is described in
Selenium's official docs.