Update README.rst
I was trying to run the browser in headless mode, but that's not explained in the project documentation, so I had to read through the code and found that you use the undetected-chromedriver project. I read their documentation to figure out how to run the browser in headless mode. I think adding this information to the project documentation will save time for other developers.